feat: refatoração e refatoração
