+++ package.json
@@ -4,7 +4,6 @@
   "description": "",
   "main": "./src/app.js",
   "scripts": {
-    "db": "cd src && mkdir DB && cd DB && touch Comands.json && echo {} > Comands.json",
     "start": "node .",
     "dev": "nodemon .",
     "test": "mocha ./src/tests/**/*.test.js",
