--- src/DB/Comands.json
@@ -1,10 +1,0 @@
-{
-  "groupPerms": [
-    "5521974915761@c.us",
-    "5521984661775-1514851884@g.us"
-  ],
-  "environment": [],
-  "userPerms": [
-    "5521988328302@c.us"
-  ]
-}
