+++ src/services/index.js
@@ -1,5 +1,5 @@
 const validateServices = require('./validateService');
 
 module.exports = {
-  ...validateServices,
+  validateServices,
 }
